Write documentation for utils.ts
