+++ Lara-JS/src-api/visualization/public/js/utils.ts
@@ -1,3 +1,10 @@
+/**
+ * @brief Counts the number of occurrences of a character in a string.
+ * 
+ * @param str String to search
+ * @param char Target character
+ * @return Number of occurrences of char in str
+ */
 const countChar = (str: string, char: string): number => {
   let count = 0;
   for (const c of str) {
@@ -7,12 +14,4 @@ const countChar = (str: string, char: string): number => {
   return count;
 }
 
-const createIcon = (name: string): HTMLElement => {
-  const icon = document.createElement('span');
-  icon.classList.add('icon', 'material-symbols-outlined');
-  icon.textContent = name;
-  
-  return icon;
-}
-
-export { countChar, createIcon };
+export { countChar };
